requirements.txt
# TempoGraph - Video Intelligence Pipeline
# Python dependencies

# Core dependencies
torch>=2.0.0
transformers>=4.35.0
pydantic>=2.0.0
networkx>=3.0.0
matplotlib>=3.7.0

# Video processing
opencv-python>=4.8.0
ffmpeg-python>=0.2.0
moviepy>=1.0.3

# VLM backends
google-genai>=0.2.0
qwen-vl-utils>=0.0.1
bitsandbytes>=0.41.0

# Object detection
ultralytics>=8.0.0

# Audio processing
openai-whisper>=20231117

# Depth estimation
depth-anything-v2>=1.0.0

# UI
streamlit>=1.28.0
plotly>=5.17.0

# Utilities
pyyaml>=6.0
huggingface-hub>=0.19.0

# Development
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0